rework delete contact

+++ services/contactsServices.js
@@ -8,12 +8,17 @@ export const getContactById = (contactId) => {
   return Contact.findByPk(contactId);
 };
 
-export const removeContact = (contactId) => {
-  return Contact.destroy({ where: { id: contactId } });
-};
-
 export const addContact = (data) => {
   return Contact.create(data);
+};
+
+export const removeContact = async (contactId) => {
+  const contact = await getContactById(contactId);
+
+  if (contact) {
+    await Contact.destroy({ where: { id: contactId } });
+    return contact;
+  } else return null;
 };
 
 export const updateContact = async (contactId, data) => {
